Abstract create
