+++ underfs/s3a/src/main/java/alluxio/underfs/s3a/S3AUnderFileSystem.java
@@ -17,7 +17,6 @@ import alluxio.Constants;
 import alluxio.PropertyKey;
 import alluxio.underfs.ObjectUnderFileSystem;
 import alluxio.underfs.UnderFileSystem;
-import alluxio.underfs.options.CreateOptions;
 import alluxio.underfs.options.DeleteOptions;
 import alluxio.underfs.options.MkdirsOptions;
 import alluxio.util.CommonUtils;
@@ -68,9 +67,6 @@ public class S3AUnderFileSystem extends ObjectUnderFileSystem {
 
   /** Suffix for an empty file to flag it as a directory. */
   private static final String FOLDER_SUFFIX = "_$folder$";
-
-  /** Value used to indicate folder structure in S3. */
-  private static final String PATH_SEPARATOR = "/";
 
   /** Static hash for a directory's empty contents. */
   private static final String DIR_HASH;
@@ -224,11 +220,8 @@ public class S3AUnderFileSystem extends ObjectUnderFileSystem {
   }
 
   @Override
-  public OutputStream createDirect(String path, CreateOptions options) throws IOException {
-    if (mkdirs(getParentKey(path), true)) {
-      return new S3AOutputStream(mBucketName, stripPrefixIfPresent(path), mManager);
-    }
-    return null;
+  protected OutputStream createOutputStream(String path) throws IOException {
+    return new S3AOutputStream(mBucketName, stripPrefixIfPresent(path), mManager);
   }
 
   @Override
@@ -594,31 +587,9 @@ public class S3AUnderFileSystem extends ObjectUnderFileSystem {
     }
   }
 
-  /**
-   * @param key the key to get the parent of
-   * @return the parent key, or null if the parent does not exist
-   */
-  private String getParentKey(String key) {
-    // Root does not have a parent.
-    if (isRoot(key)) {
-      return null;
-    }
-    int separatorIndex = key.lastIndexOf(PATH_SEPARATOR);
-    if (separatorIndex < 0) {
-      return null;
-    }
-    return key.substring(0, separatorIndex);
-  }
-
-  /**
-   * Checks if the key is the root.
-   *
-   * @param key the key to check
-   * @return true if the key is the root, false otherwise
-   */
-  private boolean isRoot(String key) {
-    return PathUtils.normalizePath(key, PATH_SEPARATOR).equals(
-        PathUtils.normalizePath(Constants.HEADER_S3A + mBucketName, PATH_SEPARATOR));
+  @Override
+  protected String getRootKey() {
+    return Constants.HEADER_S3A + mBucketName;
   }
 
   /**
